moved test_get_month into tests folder
I still need to figure out how to use the 'old tests' as a test.
